Splitted the 'Constant' module across five files
